Restructure and introduce placeholders for some of the required components
