--- api/src/main/java/org/hawkular/btm/api/model/BusinessTransactionList.java
@@ -1,83 +1,0 @@
-/*
- * Copyright 2015 Red Hat, Inc. and/or its affiliates
- * and other contributors as indicated by the @author tags.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.hawkular.btm.api.model;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-
-/**
- * This class represents a list of business transactions.
- *
- * @author gbrown
- *
- */
-public class BusinessTransactionList {
-
-    @JsonInclude
-    private List<BusinessTransaction> transactions = new ArrayList<BusinessTransaction>();
-
-    public BusinessTransactionList() {
-    }
-
-    /**
-     * @return the transactions
-     */
-    public List<BusinessTransaction> getTransactions() {
-        return transactions;
-    }
-
-    /**
-     * @param transactions the transactions to set
-     */
-    public void setTransactions(List<BusinessTransaction> transactions) {
-        this.transactions = transactions;
-    }
-
-    /* (non-Javadoc)
-     * @see java.lang.Object#hashCode()
-     */
-    @Override
-    public int hashCode() {
-        final int prime = 31;
-        int result = 1;
-        result = prime * result + ((transactions == null) ? 0 : transactions.hashCode());
-        return result;
-    }
-
-    /* (non-Javadoc)
-     * @see java.lang.Object#equals(java.lang.Object)
-     */
-    @Override
-    public boolean equals(Object obj) {
-        if (this == obj)
-            return true;
-        if (obj == null)
-            return false;
-        if (getClass() != obj.getClass())
-            return false;
-        BusinessTransactionList other = (BusinessTransactionList) obj;
-        if (transactions == null) {
-            if (other.transactions != null)
-                return false;
-        } else if (!transactions.equals(other.transactions))
-            return false;
-        return true;
-    }
-
-}
