--- api/src/main/java/org/hawkular/btm/api/model/Component.java
@@ -1,118 +1,0 @@
-/*
- * Copyright 2015 Red Hat, Inc. and/or its affiliates
- * and other contributors as indicated by the @author tags.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.hawkular.btm.api.model;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.wordnik.swagger.annotations.ApiModel;
-
-/**
- * This class represents the invocation of a component.
- *
- * @author gbrown
- *
- */
-@ApiModel(parent=InvocationNode.class)
-public class Component extends InvocationNode {
-
-    @JsonInclude
-    private String componentType;
-
-    @JsonInclude
-    private String uri;
-
-    public Component() {
-    }
-
-    public Component(String componentType, String uri) {
-        this.componentType = componentType;
-        this.uri = uri;
-    }
-
-    /**
-     * @return the componentType
-     */
-    public String getComponentType() {
-        return componentType;
-    }
-
-    /**
-     * @param componentType the componentType to set
-     */
-    public void setComponentType(String componentType) {
-        this.componentType = componentType;
-    }
-
-    /**
-     * @return the uri
-     */
-    public String getUri() {
-        return uri;
-    }
-
-    /**
-     * @param uri the uri to set
-     */
-    public void setUri(String uri) {
-        this.uri = uri;
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (this == obj) {
-            return true;
-        }
-        if (!super.equals(obj)) {
-            return false;
-        }
-        if (getClass() != obj.getClass()) {
-            return false;
-        }
-        Component other = (Component) obj;
-        if (componentType == null) {
-            if (other.componentType != null) {
-                return false;
-            }
-        } else if (!componentType.equals(other.componentType)) {
-            return false;
-        }
-        if (uri == null) {
-            if (other.uri != null) {
-                return false;
-            }
-        } else if (!uri.equals(other.uri)) {
-            return false;
-        }
-        return true;
-    }
-
-    @Override
-    public int hashCode() {
-        final int prime = 31;
-        int result = super.hashCode();
-        result = prime * result + ((componentType == null) ? 0 : componentType.hashCode());
-        result = prime * result + ((uri == null) ? 0 : uri.hashCode());
-        return result;
-    }
-
-    @Override
-    public String toString() {
-        return "ComponentInvocation{" +
-                "uri=" + (uri == null ? null : '\'' + uri + '\'') +
-                ", type='" + componentType + '\'' +
-                '}';
-    }
-}
