+++ api/src/main/java/org/hawkular/btm/api/model/Level.java
@@ -14,25 +14,34 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.btm.api.services;
-
-import java.util.List;
-
-import org.hawkular.btm.api.model.btxn.BusinessTransaction;
+package org.hawkular.btm.api.model;
 
 /**
- * This interface represents the service used to store and retrieve business
- * transactions.
+ * This enumerated type represents the level of information to be
+ * collected or presented.
  *
  * @author gbrown
  */
-public interface BusinessTransactionService {
+public enum Level {
 
     /**
-     * This method stores the list of business transactions (fragments).
-     *
-     * @param btxns The list of business transactions
+     * Only deal with communication level information.
      */
-    void store(List<BusinessTransaction> btxns);
+    Communication,
+
+    /**
+     * Service level information.
+     */
+    Service,
+
+    /**
+     * Component level information.
+     */
+    Component,
+
+    /**
+     * Detailed level information.
+     */
+    Details
 
 }
