--- api/src/main/java/org/hawkular/btm/api/model/Service.java
@@ -1,113 +1,0 @@
-/*
- * Copyright 2015 Red Hat, Inc. and/or its affiliates
- * and other contributors as indicated by the @author tags.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.hawkular.btm.api.model;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.wordnik.swagger.annotations.ApiModel;
-
-/**
- * This class represents a service invocation.
- *
- * @author gbrown
- *
- */
-@ApiModel(parent=InvocationNode.class)
-public class Service extends InvocationNode {
-
-    @JsonInclude
-    private String serviceType;
-
-    @JsonInclude
-    private String operation;
-
-    public Service() {
-    }
-
-    public Service(String serviceType, String operation) {
-        this.serviceType = serviceType;
-        this.operation = operation;
-    }
-
-    /**
-     * @return the serviceType
-     */
-    public String getServiceType() {
-        return serviceType;
-    }
-
-    /**
-     * @param serviceType the serviceType to set
-     */
-    public void setServiceType(String serviceType) {
-        this.serviceType = serviceType;
-    }
-
-    /**
-     * @return the operation
-     */
-    public String getOperation() {
-        return operation;
-    }
-
-    /**
-     * @param operation the operation to set
-     */
-    public void setOperation(String operation) {
-        this.operation = operation;
-    }
-
-    @Override
-    public int hashCode() {
-        final int prime = 31;
-        int result = super.hashCode();
-        result = prime * result
-                + ((operation == null) ? 0 : operation.hashCode());
-        result = prime * result
-                + ((serviceType == null) ? 0 : serviceType.hashCode());
-        return result;
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-        if (this == obj) {
-            return true;
-        }
-        if (!super.equals(obj)) {
-            return false;
-        }
-        if (getClass() != obj.getClass()) {
-            return false;
-        }
-        Service other = (Service) obj;
-        if (operation == null) {
-            if (other.operation != null) {
-                return false;
-            }
-        } else if (!operation.equals(other.operation)) {
-            return false;
-        }
-        if (serviceType == null) {
-            if (other.serviceType != null) {
-                return false;
-            }
-        } else if (!serviceType.equals(other.serviceType)) {
-            return false;
-        }
-        return true;
-    }
-
-}
