+++ api/src/main/java/org/hawkular/btm/api/model/btxn/Consumer.java
@@ -1,0 +1,112 @@
+/*
+ * Copyright 2015 Red Hat, Inc. and/or its affiliates
+ * and other contributors as indicated by the @author tags.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.hawkular.btm.api.model.btxn;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.wordnik.swagger.annotations.ApiModel;
+
+/**
+ * This class represents the situation where the business transaction flow
+ * is triggered by a communication from an external participant.
+ *
+ * @author gbrown
+ *
+ */
+@ApiModel(parent=InvocationNode.class)
+public class Consumer extends InvocationNode {
+
+    @JsonInclude
+    private String endpointType;
+
+    @JsonInclude
+    private String uri;
+
+    public Consumer() {
+    }
+
+    public Consumer(String endpointType, String uri) {
+        this.endpointType = endpointType;
+        this.uri = uri;
+    }
+
+    /**
+     * @return the endpointType
+     */
+    public String getEndpointType() {
+        return endpointType;
+    }
+
+    /**
+     * @param endpointType the endpointType to set
+     */
+    public void setEndpointType(String endpointType) {
+        this.endpointType = endpointType;
+    }
+
+    /**
+     * @return the uri
+     */
+    public String getUri() {
+        return uri;
+    }
+
+    /**
+     * @param uri the uri to set
+     */
+    public void setUri(String uri) {
+        this.uri = uri;
+    }
+
+    @Override
+    public int hashCode() {
+        final int prime = 31;
+        int result = super.hashCode();
+        result = prime * result + ((endpointType == null) ? 0 : endpointType.hashCode());
+        result = prime * result + ((uri == null) ? 0 : uri.hashCode());
+        return result;
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj) {
+            return true;
+        }
+        if (!super.equals(obj)) {
+            return false;
+        }
+        if (getClass() != obj.getClass()) {
+            return false;
+        }
+        Consumer other = (Consumer) obj;
+        if (endpointType == null) {
+            if (other.endpointType != null) {
+                return false;
+            }
+        } else if (!endpointType.equals(other.endpointType)) {
+            return false;
+        }
+        if (uri == null) {
+            if (other.uri != null) {
+                return false;
+            }
+        } else if (!uri.equals(other.uri)) {
+            return false;
+        }
+        return true;
+    }
+
+}
