+++ api/src/main/java/org/hawkular/btm/api/model/btxn/CorrelationIdentifier.java
@@ -1,0 +1,155 @@
+/*
+ * Copyright 2015 Red Hat, Inc. and/or its affiliates
+ * and other contributors as indicated by the @author tags.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.hawkular.btm.api.model.btxn;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonInclude.Include;
+
+/**
+ * This class represents a Correlation Identifier. These identifies are used to
+ * link information generated by multiple systems involved in the execution of
+ * a business transaction instance.
+ *
+ * @author gbrown
+ *
+ */
+public class CorrelationIdentifier {
+
+    @JsonInclude
+    private String value;
+
+    @JsonInclude
+    private Scope scope;
+
+    @JsonInclude(Include.NON_DEFAULT)
+    private long expiration = 0;
+
+    public CorrelationIdentifier() {
+    }
+
+    public CorrelationIdentifier(String value, Scope scope, long expiration) {
+        this.value = value;
+        this.scope = scope;
+        this.expiration = expiration;
+    }
+
+    /**
+     * @return the value
+     */
+    public String getValue() {
+        return value;
+    }
+
+    /**
+     * @param value the value to set
+     */
+    public void setValue(String value) {
+        this.value = value;
+    }
+
+    /**
+     * @return the scope
+     */
+    public Scope getScope() {
+        return scope;
+    }
+
+    /**
+     * @param scope the scope to set
+     */
+    public void setScope(Scope scope) {
+        this.scope = scope;
+    }
+
+    /**
+     * @return the expiration
+     */
+    public long getExpiration() {
+        return expiration;
+    }
+
+    /**
+     * @param expiration the expiration to set
+     */
+    public void setExpiration(long expiration) {
+        this.expiration = expiration;
+    }
+
+    @Override
+    public int hashCode() {
+        final int prime = 31;
+        int result = 1;
+        result = prime * result + (int) (expiration ^ (expiration >>> 32));
+        result = prime * result + ((scope == null) ? 0 : scope.hashCode());
+        result = prime * result + ((value == null) ? 0 : value.hashCode());
+        return result;
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj) {
+            return true;
+        }
+        if (obj == null) {
+            return false;
+        }
+        if (getClass() != obj.getClass()) {
+            return false;
+        }
+        CorrelationIdentifier other = (CorrelationIdentifier) obj;
+        if (expiration != other.expiration) {
+            return false;
+        }
+        if (scope != other.scope) {
+            return false;
+        }
+        if (value == null) {
+            if (other.value != null) {
+                return false;
+            }
+        } else if (!value.equals(other.value)) {
+            return false;
+        }
+        return true;
+    }
+
+    /**
+     * This enumerated value represents the scope of the correlation identifier.
+     *
+     * @author gbrown
+     *
+     */
+    public enum Scope {
+        /**
+         * The 'Global' scope means the identifier is unique over the end to end
+         * life of the business transaction.
+         */
+        Global,
+
+        /**
+         * The 'Local' scope means the identifier is unique within an island
+         * of the business transaction (e.g. a business process instance).
+         */
+        Local,
+
+        /**
+         * The 'Exchange' scope means the identifier is only unique in relation
+         * to a specific exchange between two (or more) participants.
+         */
+        Exchange
+    }
+}
