+++ api/src/main/java/org/hawkular/btm/api/repositories/BusinessTransactionRepository.java
@@ -14,25 +14,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hawkular.btm.api.services;
-
-import java.util.List;
-
-import org.hawkular.btm.api.model.btxn.BusinessTransaction;
+package org.hawkular.btm.api.repositories;
 
 /**
- * This interface represents the service used to store and retrieve business
- * transactions.
- *
  * @author gbrown
  */
-public interface BusinessTransactionService {
-
-    /**
-     * This method stores the list of business transactions (fragments).
-     *
-     * @param btxns The list of business transactions
-     */
-    void store(List<BusinessTransaction> btxns);
+public interface BusinessTransactionRepository {
 
 }
