+++ backend/rest/src/main/java/org/hawkular/btm/rest/ApiError.java
