+++ backend/rest/src/main/java/org/hawkular/btm/rest/swagger/JaxRsFilter.java
