+++ backend/rest/src/main/webapp/WEB-INF/beans.xml
