+++ backend/rest/src/main/webapp/WEB-INF/jboss-deployment-structure.xml
